fixup! bitmap type

+++ bitmap.h
@@ -20,36 +20,37 @@
 
 #ifndef BITMAP_H_
 #define BITMAP_H_
+#include <stddef.h>
 
 /* Test a bit, 0 = not set, 1 = set */
-static inline unsigned char test(unsigned char *bm, int ndx) {
+static inline unsigned char test(unsigned char *bm, size_t ndx) {
   return ( bm[ndx>>3] & (1 << (ndx&0x07))?1:0 ); 
 }
 
 /* Clear a bit (set it to 0) */
-static inline void clear(unsigned char *bm, int ndx) {
+static inline void clear(unsigned char *bm, size_t ndx) {
   bm[ndx>>3] &= ~(1 << (ndx&0x07));
 }
 
 /* Set a bit to 1 */
-static inline void set(unsigned char *bm, int ndx) {
+static inline void set(unsigned char *bm, size_t ndx) {
   bm[ndx>>3] |= (1 << (ndx&0x07));
 }
 
 /* Flip a bit's value 0->1 or 1->0 */
-static inline void toggle(unsigned char *bm, int ndx) {
+static inline void toggle(unsigned char *bm, size_t ndx) {
   bm[ndx>>3] ^= (1 << (ndx&0x07)); 
 }
 
 /* Get two bits */
-static inline unsigned char nuc(unsigned char *bm, int ndx) {
+static inline unsigned char nuc(unsigned char *bm, size_t ndx) {
   ndx *= 2;
   return ((bm[ndx >> 3] >> (ndx & 0x07)) & 0x03);
 }
 
 /* Get six bits, which can be in two bytes
    out-of-bound read may happen but will not affect result */
-static inline unsigned char trinuc(unsigned char *bm, int ndx) {
+static inline unsigned char trinuc(unsigned char *bm, size_t ndx) {
   ndx *= 2;
   unsigned byte = ndx >> 3;
   unsigned stretch = (bm[byte + 1] << 8) | bm[byte];
